+++ src/SampleTest/AspNetCore.Http/UriHelperTest.cs
@@ -1,0 +1,32 @@
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.Extensions;
+
+namespace SampleTest.AspNetCore.Http;
+
+public class UriHelperTest {
+	[Fact]
+	public void BuildAbsolute_とりあえず使ってみる() {
+		// Arrange
+		// Act
+		var url = UriHelper.BuildAbsolute(
+			"https",
+			new HostString("example.jp"),
+			pathBase: "/app",
+			path: "/default/index");
+
+		// Assert
+		Assert.Equal("https://example.jp/app/default/index", url);
+	}
+
+	[Fact]
+	public void BuildRelative_とりあえず使ってみる() {
+		// Arrange
+		// Act
+		var url = UriHelper.BuildRelative(
+			pathBase: "/app",
+			path: "/default/index");
+
+		// Assert
+		Assert.Equal("/app/default/index", url);
+	}
+}
